Congratulationsgit add *

+++ ft_strlen.c
@@ -6,7 +6,7 @@
 /*   By: cportuon <cportuon@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/08/31 11:18:06 by cportuon          #+#    #+#             */
-/*   Updated: 2022/09/08 12:07:07 by cportuon         ###   ########.fr       */
+/*   Updated: 2022/09/30 17:41:43 by cportuon         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 size_t	ft_strlen(const	char *str)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (str[i] != '\0')
